requirements.txt

cassandra-driver==3.25.0
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==1.10.3
geomet==0.2.1.post1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.16.0
threadpoolctl==2.2.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.5.0
gunicorn==20.1.0
